pyproject.toml
[tool.poetry]
name = "signature_checker"
version = "0.1.0"
description = ""
authors = ["yury labovich"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
pycryptodome = "^3.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
